flyway

+++ eureka-client-user/src/main/java/com/guier/eurekaclientuser/controller/UserController.java
@@ -26,7 +26,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public Optional<User> findById(@PathVariable Long id) {
+    public Optional<User> findById(@PathVariable Integer id) {
         System.out.println(port);
         return this.userRepository.findById(id);
     }
